Add variant in dropdown menu
