+++ src/stories/DropdownMenu/styles.css
@@ -1,0 +1,16 @@
+.dropdown-menu {
+    section {
+        padding: var(--dui-spacing-medium);
+
+        .content {
+            display: flex;
+            flex-wrap: wrap;
+
+            >* {
+                flex-shrink: 0;
+                margin: var(--dui-spacing-large);
+            }
+        }
+
+    }
+}
